# apps/server/openapi.v2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ashyq Bilim API",
    "description": "University LMS / MOOC platform API (v2, Rust rewrite).",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/v2/activities/{id}": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "Full activity including content/details/settings.",
        "operationId": "get_activity",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Activity id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ActivityId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Activity detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityDetail"
                }
              }
            }
          },
          "404": {
            "description": "Unknown or inaccessible",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "courses"
        ],
        "summary": "Delete an activity; chapter siblings renumber to stay contiguous.",
        "operationId": "delete_activity",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Activity id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ActivityId"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "403": {
            "description": "No write access",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "courses"
        ],
        "summary": "Partial update: name, publish state, content/details/settings, or the\ntype pair (both `activity_type` and `activity_sub_type` together).",
        "operationId": "update_activity",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Activity id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ActivityId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateActivityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityDetail"
                }
              }
            }
          },
          "403": {
            "description": "No write access",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/activities/{id}/blocks": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "Blocks attached to an activity.",
        "operationId": "list_blocks",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Activity id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ActivityId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Blocks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Block"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Unknown or inaccessible",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "courses"
        ],
        "summary": "Attach a file block to an activity by claiming a finalized upload whose\npurpose matches the block type (`image`→`block-image`, etc.).",
        "operationId": "create_block",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Activity id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ActivityId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBlockRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Block"
                }
              }
            }
          },
          "403": {
            "description": "No write access / not your upload",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Upload not finalized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Bad block type or wrong upload purpose",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/activities/{id}/move": {
      "post": {
        "tags": [
          "courses"
        ],
        "summary": "Move an activity within its chapter, or to another chapter of the same\ncourse via `chapter_id`.",
        "operationId": "move_activity",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Activity id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ActivityId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveActivityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Moved"
          },
          "403": {
            "description": "No write access",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Cross-course move",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/blocks/{id}": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "One block (visibility follows the course).",
        "operationId": "get_block",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Block id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/BlockId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Block",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Block"
                }
              }
            }
          },
          "404": {
            "description": "Unknown or inaccessible",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "courses"
        ],
        "summary": "Delete a block and release its stored file for reaping.",
        "operationId": "delete_block",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Block id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/BlockId"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "403": {
            "description": "No write access",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/chapters/{id}": {
      "delete": {
        "tags": [
          "courses"
        ],
        "summary": "Delete a chapter and its activities; siblings renumber to stay contiguous.",
        "operationId": "delete_chapter",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Chapter id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ChapterId"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "403": {
            "description": "No write access",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "courses"
        ],
        "summary": "Rename/redescribe a chapter.",
        "operationId": "update_chapter",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Chapter id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ChapterId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateChapterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Chapter"
                }
              }
            }
          },
          "403": {
            "description": "No write access",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/chapters/{id}/activities": {
      "post": {
        "tags": [
          "courses"
        ],
        "summary": "Append an activity to a chapter (type/subtype pair must be valid).",
        "operationId": "create_activity",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Chapter id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ChapterId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateActivityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created (appended last)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Activity"
                }
              }
            }
          },
          "403": {
            "description": "No write access",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Invalid type/subtype pair",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/chapters/{id}/move": {
      "post": {
        "tags": [
          "courses"
        ],
        "summary": "Move a chapter to a new position (clamped; siblings renumber).",
        "operationId": "move_chapter",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Chapter id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ChapterId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveChapterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Moved"
          },
          "403": {
            "description": "No write access",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/courses/{id}/chapters": {
      "post": {
        "tags": [
          "courses"
        ],
        "summary": "Append a chapter to a course (course write access).",
        "operationId": "create_chapter",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Course id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CourseId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChapterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created (appended last)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Chapter"
                }
              }
            }
          },
          "403": {
            "description": "No write access",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/courses/{id}/curriculum": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "Chapters with nested activities, in course order.",
        "operationId": "get_curriculum",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Course id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CourseId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ordered chapters + activities",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Curriculum"
                }
              }
            }
          },
          "404": {
            "description": "Unknown or inaccessible course",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActivityDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Activity"
          },
          {
            "type": "object",
            "required": [
              "content",
              "details",
              "settings"
            ],
            "properties": {
              "content": {
                "description": "Editor content (dynamic pages) or type-specific payload."
              },
              "details": {},
              "settings": {}
            }
          }
        ],
        "description": "Full activity view with the heavy jsonb columns."
      },
      "ActivityId": {
        "type": "string",
        "format": "uuid"
      },
...
      "Block": {
        "type": "object",
        "required": [
          "id",
          "activity_id",
          "block_type",
          "content",
          "created_at_unix"
        ],
        "properties": {
          "activity_id": {
            "$ref": "#/components/schemas/ActivityId"
          },
          "block_type": {
            "type": "string",
            "description": "`image`, `pdf`, `video` (or `custom` for migrated legacy rows)."
          },
          "content": {
            "description": "`{upload_id, file_key, file_name, file_size, file_type}`."
          },
          "created_at_unix": {
            "type": "integer",
            "format": "int64"
          },
          "id": {
            "$ref": "#/components/schemas/BlockId"
          }
        }
      },
...
      "CreateBlockRequest": {
        "type": "object",
        "required": [
          "block_type",
          "upload_id"
        ],
        "properties": {
          "block_type": {
            "type": "string",
            "description": "`image`, `pdf`, or `video`."
          },
          "file_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Original client-side file name, for display."
          },
          "upload_id": {
            "type": "string",
            "format": "uuid",
            "description": "A finalized upload (purpose must match the block type)."
          }
        },
        "additionalProperties": false
      },
...
      "MoveActivityRequest": {
        "type": "object",
        "description": "Target slot for an activity move; `chapter_id` reparents within the same\ncourse, omitted keeps the current chapter.",
        "required": [
          "position"
        ],
        "properties": {
          "chapter_id": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ChapterId"
              }
            ]
          },
          "position": {
            "type": "integer",
            "format": "int32",
            "description": "1-based target position (in the destination chapter)."
          }
        },
        "additionalProperties": false
      },
...
      "Problem": {
        "type": "object",
        "description": "The wire shape of every error response.",
        "required": [
          "type",
          "status",
          "code",
          "title"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode",
            "description": "Stable machine code — the frontend's i18n key."
          },
          "detail": {
            "type": [
              "string",
              "null"
            ]
          },
          "field_errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          },
          "request_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Correlation id; also present as the `x-request-id` response header."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "description": "Stable documentation URI for this error class."
          }
        }
      },
...
      "UpdateActivityRequest": {
        "type": "object",
        "properties": {
          "activity_sub_type": {
            "type": [
              "string",
              "null"
            ]
          },
          "activity_type": {
            "type": [
              "string",
              "null"
            ],
            "description": "Change together with `activity_sub_type` (both or neither)."
          },
          "content": {},
          "details": {},
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "published": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "settings": {}
        },
        "additionalProperties": false
      },
...
    }
  },
...
}